# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для вычисления времени выполнения SQL-запросов",
        "title": "Query Execution Time Calculation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/indexes_query/{index_id}/{query_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет параметры индекса в конкретном запросе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indexes-query"
                ],
                "summary": "Изменить данные индекса в запросе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID индекса",
                        "name": "index_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID запроса",
                        "name": "query_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.IndexesQueryJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "$ref": "#/definitions/apitypes.IndexesQueryJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь индекса и запроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "indexes-query"
                ],
                "summary": "Удалить индекс из запроса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID индекса",
                        "name": "index_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID запроса",
                        "name": "query_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный запрос",
                        "schema": {
                            "$ref": "#/definitions/apitypes.QueryJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apitypes.IndexesQueryJSON": {
            "type": "object",
            "properties": {
                "cardinality": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "index_id": {
                    "type": "integer"
                },
                "query_id": {
                    "type": "integer"
                },
                "recieved_rows": {
                    "type": "integer"
                },
                "rows_count": {
                    "type": "integer"
                },
                "table_field": {
                    "type": "string"
                }
            }
        },
        "apitypes.QueryJSON": {
            "type": "object",
            "properties": {
                "creator_login": {
                    "type": "string"
                },
                "date_create": {
                    "type": "string"
                },
                "date_finish": {
                    "type": "string"
                },
                "date_form": {
                    "type": "string"
                },
                "date_query": {
                    "type": "string"
                },
                "execution_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}